requirements.txt
asgiref==3.7.2
Django==4.2.7
django-environ==0.11.2
django-simple-history==3.4.0
psycopg2==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
